+++ lib/group_select.dart
@@ -1,4 +1,4 @@
 library group_select;
 
 export 'components/group_select.dart';
-export 'controllers/group_select/group_select_controller.dart';
+export 'controllers/select_group_controller.dart';